requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.56.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
openai==1.99.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
